+++ packages/provider-github/src/lib/getGitRemote.ts
@@ -1,7 +1,4 @@
-import cacheManager from './cacheManager.js';
-import logger from './logger.js';
-import { promptSelect } from './prompts.js';
-import { spawn } from './spawn.js';
+import { cacheManager, promptSelect, spawn } from '@rnef/tools';
 
 export async function getGitRemote() {
   let gitRemote = cacheManager.get('gitRemote');
@@ -26,8 +23,6 @@ export async function getGitRemote() {
   } else if (remotes.length === 1) {
     gitRemote = remotes[0];
   } else {
-    // @todo add "learn more" link to docs when available
-    logger.warn('No git remote found. Proceeding with local build.');
     return null;
   }