+++ packages/tools/src/index.ts
@@ -7,16 +7,15 @@ export { default as cacheManager } from './lib/cacheManager.js';
 export * from './lib/parse-args.js';
 export * from './lib/path.js';
 export * from './lib/project.js';
-
 export * from './lib/build-cache/common.js';
 export * from './lib/build-cache/localBuildCache.js';
-export * from './lib/build-cache/remoteBuildCache.js';
-export * from './lib/build-cache/github/config.js';
-export * from './lib/build-cache/github/GitHubBuildCache.js';
 export { findDevServerPort } from './lib/dev-server/findDevServerPort.js';
 export { isDevServerRunning } from './lib/dev-server/isDevServerRunning.js';
 export { isInteractive } from './lib/isInteractive.js';
 export { spawn, SubprocessError } from './lib/spawn.js';
 export { color } from './lib/color.js';
 export { runHermes } from './lib/hermes.js';
-export { fetchCachedBuild } from './lib/build-cache/fetchCachedBuild.js';
+export {
+  fetchCachedBuild,
+  handleDownloadResponse,
+} from './lib/build-cache/fetchCachedBuild.js';