+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,66 @@
 # @rnef/cli
 
+## 0.7.13
+
+### Patch Changes
+
+- @rnef/config@0.7.13
+- @rnef/tools@0.7.13
+
+## 0.7.12
+
+### Patch Changes
+
+- Updated dependencies [dfed056]
+  - @rnef/config@0.7.12
+  - @rnef/tools@0.7.12
+
+## 0.7.11
+
+### Patch Changes
+
+- Updated dependencies [dc7ba26]
+  - @rnef/tools@0.7.11
+  - @rnef/config@0.7.11
+
+## 0.7.10
+
+### Patch Changes
+
+- 3a88d7b: feat: add --all flag to list and delete, --all-but-latest to delete
+- Updated dependencies [3a88d7b]
+  - @rnef/tools@0.7.10
+  - @rnef/config@0.7.10
+
+## 0.7.9
+
+### Patch Changes
+
+- 51316ba: fix: remote-cache list returninig undefined instead of empty
+- fc771e7: fix: download message in remote-cache --json
+- Updated dependencies [fc771e7]
+  - @rnef/tools@0.7.9
+  - @rnef/config@0.7.9
+
+## 0.7.8
+
+### Patch Changes
+
+- af32d6f: feat: add remote-cache command
+- e8c9da7: refactor: internal plugins
+- Updated dependencies [184cc74]
+- Updated dependencies [af32d6f]
+- Updated dependencies [e8c9da7]
+  - @rnef/config@0.7.8
+  - @rnef/tools@0.7.8
+
+## 0.7.7
+
+### Patch Changes
+
+- @rnef/config@0.7.7
+- @rnef/tools@0.7.7
+
 ## 0.7.6
 
 ### Patch Changes