+++ packages/platform-android/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @rnef/platform-android
 
+## 0.7.13
+
+### Patch Changes
+
+- 9fcbbee: feat: ask whether to continue with local build when remote fails
+  - @rnef/tools@0.7.13
+
+## 0.7.12
+
+### Patch Changes
+
+- @rnef/tools@0.7.12
+
+## 0.7.11
+
+### Patch Changes
+
+- Updated dependencies [dc7ba26]
+  - @rnef/tools@0.7.11
+
+## 0.7.10
+
+### Patch Changes
+
+- Updated dependencies [3a88d7b]
+  - @rnef/tools@0.7.10
+
+## 0.7.9
+
+### Patch Changes
+
+- Updated dependencies [fc771e7]
+  - @rnef/tools@0.7.9
+
+## 0.7.8
+
+### Patch Changes
+
+- Updated dependencies [184cc74]
+- Updated dependencies [af32d6f]
+  - @rnef/tools@0.7.8
+
+## 0.7.7
+
+### Patch Changes
+
+- 5944f07: feat: retry installing android app when insufficient storage
+- 07d7668: fix: lazy load autolinking platform config
+  - @rnef/tools@0.7.7
+
 ## 0.7.6
 
 ### Patch Changes