+++ packages/platform-apple-helpers/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @rnef/platform-apple-helpers
 
+## 0.7.13
+
+### Patch Changes
+
+- 9fcbbee: feat: ask whether to continue with local build when remote fails
+- 80cc3d1: unify --destination and --destinations flags; fix building universal (device+simu) iOS brownfield framework
+  - @rnef/tools@0.7.13
+
+## 0.7.12
+
+### Patch Changes
+
+- @rnef/tools@0.7.12
+
+## 0.7.11
+
+### Patch Changes
+
+- 34f1c59: Interactive Xcode scheme selector for `package:ios` command
+- Updated dependencies [dc7ba26]
+  - @rnef/tools@0.7.11
+
+## 0.7.10
+
+### Patch Changes
+
+- Updated dependencies [3a88d7b]
+  - @rnef/tools@0.7.10
+
+## 0.7.9
+
+### Patch Changes
+
+- Updated dependencies [fc771e7]
+  - @rnef/tools@0.7.9
+
+## 0.7.8
+
+### Patch Changes
+
+- Updated dependencies [184cc74]
+- Updated dependencies [af32d6f]
+  - @rnef/tools@0.7.8
+
+## 0.7.7
+
+### Patch Changes
+
+- @rnef/tools@0.7.7
+
 ## 0.7.6
 
 ### Patch Changes