+++ packages/plugin-brownfield-ios/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @rnef/plugin-brownfield-ios
 
+## 0.7.13
+
+### Patch Changes
+
+- 80cc3d1: unify --destination and --destinations flags; fix building universal (device+simu) iOS brownfield framework
+- Updated dependencies [9fcbbee]
+- Updated dependencies [80cc3d1]
+  - @rnef/platform-apple-helpers@0.7.13
+  - @rnef/tools@0.7.13
+
+## 0.7.12
+
+### Patch Changes
+
+- @rnef/platform-apple-helpers@0.7.12
+- @rnef/tools@0.7.12
+
+## 0.7.11
+
+### Patch Changes
+
+- 34f1c59: Interactive Xcode scheme selector for `package:ios` command
+- Updated dependencies [dc7ba26]
+- Updated dependencies [34f1c59]
+  - @rnef/tools@0.7.11
+  - @rnef/platform-apple-helpers@0.7.11
+
+## 0.7.10
+
+### Patch Changes
+
+- Updated dependencies [3a88d7b]
+  - @rnef/tools@0.7.10
+  - @rnef/platform-apple-helpers@0.7.10
+
+## 0.7.9
+
+### Patch Changes
+
+- Updated dependencies [fc771e7]
+  - @rnef/tools@0.7.9
+  - @rnef/platform-apple-helpers@0.7.9
+
+## 0.7.8
+
+### Patch Changes
+
+- Updated dependencies [184cc74]
+- Updated dependencies [af32d6f]
+  - @rnef/tools@0.7.8
+  - @rnef/platform-apple-helpers@0.7.8
+
+## 0.7.7
+
+### Patch Changes
+
+- 07d7668: fix: lazy load autolinking platform config
+  - @rnef/platform-apple-helpers@0.7.7
+  - @rnef/tools@0.7.7
+
 ## 0.7.6
 
 ### Patch Changes