+++ packages/tools/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @rnef/tools
 
+## 0.7.13
+
+## 0.7.12
+
+## 0.7.11
+
+### Patch Changes
+
+- dc7ba26: feat: extract @rnef/provider-github package
+
+## 0.7.10
+
+### Patch Changes
+
+- 3a88d7b: feat: add --all flag to list and delete, --all-but-latest to delete
+
+## 0.7.9
+
+### Patch Changes
+
+- fc771e7: fix: download message in remote-cache --json
+
+## 0.7.8
+
+### Patch Changes
+
+- 184cc74: feat: official support for custom remote cache providers
+- af32d6f: feat: add remote-cache command
+
+## 0.7.7
+
 ## 0.7.6
 
 ## 0.7.5