+++ website/CHANGELOG.md
@@ -1,5 +1,26 @@
 # rnef-docs
 
+## 1.0.5
+
+### Patch Changes
+
+- 80cc3d1: unify --destination and --destinations flags; fix building universal (device+simu) iOS brownfield framework
+
+## 1.0.4
+
+### Patch Changes
+
+- aa54bef: docs: add `llms.txt` & `llms-full.txt`
+- 0710ad9: chore: add Sitemap plugin to doc site
+- 184cc74: feat: official support for custom remote cache providers
+- ffe0cbd: chore: upgrade rspress to 2.0.0-beta.6
+
+## 1.0.3
+
+### Patch Changes
+
+- 3428485: chore: add kebab-case aliases for start command
+
 ## 1.0.2
 
 ### Patch Changes